# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Event Ticketing - CSE341 Final Project\"",
    "description": "API for ticketing data (Events, Venues, Customers, Tickets)",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "Get all events",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "description": "Create a new events",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Event data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/tickets/": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "description": "Get all tickets",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Tickets"
        ],
        "description": "Create a new ticket",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Ticket data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/venues/": {
      "get": {
        "tags": [
          "Venues"
        ],
        "description": "Get all venues",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Venues"
        ],
        "description": "Create a new venue",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Venue data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Venue"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/customers/": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "Get all customers",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "description": "Create a new customer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Customer data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Venue": {
      "type": "object",
      "properties": {
        "venueName": {
          "type": "string",
          "example": "Madison Square Garden"
        },
        "city": {
          "type": "string",
          "example": "New York"
        },
        "country": {
          "type": "string",
          "example": "United States"
        },
        "address": {
          "type": "string",
          "example": "4 Pennsylvania Plaza"
        },
        "gpsCoordinates": {
          "type": "string",
          "example": "40.7505,-73.9934"
        }
      }
    },
    "Customer": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Smith"
        },
        "email": {
          "type": "string",
          "example": "john.smith@email.com"
        },
        "phoneNumber": {
          "type": "string",
          "example": "+1-555-0101"
        },
        "gender": {
          "type": "string",
          "example": "M"
        }
      }
    },
    "Event": {
      "type": "object",
      "properties": {
        "eventName": {
          "type": "string",
          "example": "Rock Concert"
        },
        "startDate": {
          "type": "string",
          "example": "2025-09-01"
        },
        "endDate": {
          "type": "string",
          "example": "2025-09-01"
        },
        "startTime": {
          "type": "string",
          "example": "20:00"
        },
        "endTime": {
          "type": "string",
          "example": "23:00"
        },
        "capacity": {
          "type": "number",
          "example": 5000
        },
        "eventType": {
          "type": "string",
          "example": "Concert"
        },
        "eventPrice": {
          "type": "number",
          "example": 120
        },
        "venueId": {
          "type": "string",
          "example": "6890b8efd801e1a85b254a29"
        }
      }
    },
    "Ticket": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string",
          "example": "6890b8efd801e1a85b254a2f"
        },
        "eventId": {
          "type": "string",
          "example": "68940c4315d95a5bac729f04"
        },
        "ticketStatus": {
          "type": "string",
          "example": "active"
        },
        "amountPaid": {
          "type": "number",
          "example": 120
        },
        "purchaseDate": {
          "type": "string",
          "example": "2025-08-01T10:00:00Z"
        },
        "paymentMethod": {
          "type": "string",
          "example": "Credit Card"
        }
      }
    }
  }
}